requirements.txt
jira==2.0.0
arrow==0.17.0
google-api-python-client==1.12.8
google-auth-oauthlib==0.4.2
colorama==0.4.4
